requirements.txt
alembic==1.13.2
aniso8601==9.0.1
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Faker==28.4.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.1
greenlet==3.0.3
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
numpy==2.1.1
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
psycopg2==2.9.9
PyJWT==2.9.0
pytest==8.3.2
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
WTForms==3.1.2
